+++ windowsspecs/unsupported.go
@@ -1,0 +1,3 @@
+// +build !windows
+
+package specs
